Footnote Changes
Previously, UNLV required footnote text to be normal size.  The Graduate College has since removed that requirement.  I've modified this so that the footnotes are in 10pt font and there is only a single space between each footnote- This makes for a more compact footnote section in the event you have multiple, short footnotes.